+++ libraries/ESP8266WebServer/src/ESP8266WebServer-impl.h
@@ -515,7 +515,7 @@ void ESP8266WebServerTemplate<ServerType>::sendContent(Stream* content, ssize_t 
   ssize_t sent = content->sendSize(&_currentClient, content_length);
   if (sent != content_length)
   {
-    DBGWS("HTTPServer: error: short send after timeout (%d<%d)\n", sent, content_length);
+    DBGWS("HTTPServer: error: short send after timeout (%d<%d)\n", (int)sent, content_length);
   }
   if(_chunked) {
     _currentClient.printf_P(PSTR("\r\n"));
